working on PassDAO

+++ BuildWeek01_be/src/models/Pass.java
@@ -32,7 +32,7 @@ public abstract class Pass{
         this.price = price;
     }
 
-    public Pass(int id, String name, String description, double price) {
+    public Pass(long id, String name, String description, double price) {
         this.id = id;
         this.name = name;
         this.description = description;
